requirements.txt
asgiref==3.5.2
attrs==22.1.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
coreapi==2.3.3
coreschema==0.0.4
crashtest==0.3.1
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.3
django-debug-toolbar==3.7.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dulwich==0.20.50
filelock==3.8.0
html5lib==1.1
idna==3.4
importlib-metadata==5.0.0
itypes==1.2.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
jsonschema==4.17.0
keyring==23.11.0
lockfile==0.12.2
MarkupSafe==2.1.1
more-itertools==9.0.0
msgpack==1.0.4
oauthlib==3.2.2
packaging==21.3
pexpect==4.8.0
pipenv==2022.11.11
pkginfo==1.8.3
platformdirs==2.5.4
poetry==1.2.2
poetry-core==1.3.2
poetry-plugin-export==1.2.0
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.6.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.19.2
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tomlkit==0.11.6
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.7
virtualenv-clone==0.5.7
webencodings==0.5.1
zipp==3.10.0
